requirements.txt
paddlenlp==2.5.2
ppdiffusers==0.14.0
safetensors
ftfy
regex
Pillow
OmegaConf
paddlehub